requirements.txt
joblib==0.17.0
numpy==1.18.5
psycopg2==2.8.6
scikit-learn==0.23.2
scipy==1.5.4
threadpoolctl==2.1.0
matplotlib==3.3.2
pandas==1.1.3
pandas-compat==0.1.1
sshtunnel==0.4.0
Keras==2.4.3
tensorflow==2.3.1
xlwt==1.3.0
python-dotenv